Updating the project to the final version 10.0. The update is currently in progress, with 80% already completed.
In this update, the entire system has been completely overhauled, addressing errors and optimizing performance issues. While there is still much to be fixed, a significant portion of issues related to functionality, aesthetics, and design has been restructured. We continue to work on the next system update to enhance the overall product experience.